requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
chromadb==0.5.0
umap-learn==0.5.6
hdbscan==0.8.33
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.2
tqdm==4.66.4
scipy==1.13.0
torch==2.2.2
transformers==4.40.2
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
